moved LanguageChecker
